Split index.js over multiple files
