+++ src/__tests__/EventService.test.js
@@ -1,0 +1,8 @@
+import EventService from '../EventService';
+
+// TODO: Test properly
+describe('EventService.js', () => {
+  it('doesn\'t crash', () => {
+    new EventService();
+  });
+});
