+++ src/__tests__/PopulateService.test.js
@@ -1,0 +1,8 @@
+import PopulateService from '../PopulateService';
+
+// TODO: Test properly
+describe('PopulateService.js', () => {
+  it('doesn\'t crash', () => {
+    new PopulateService({}, {}, 'test');
+  });
+});
