+++ src/__tests__/Watcher.test.js
@@ -1,0 +1,8 @@
+import Watcher from '../Watcher';
+
+// TODO: Test properly
+describe('Watcher.js', () => {
+  it('doesn\'t crash', () => {
+    new Watcher({}, {}, {}, 'test');
+  });
+});
